Refactored SyncServer to SyncServerModule
