+++ pype/modules/sync_server/__init__.py
@@ -1,5 +1,5 @@
-from pype.modules.sync_server.sync_server import SyncServer
+from pype.modules.sync_server.sync_server_module import SyncServerModule
 
 
 def tray_init(tray_widget, main_widget):
-    return SyncServer()
+    return SyncServerModule()
